+++ docs/Changelog.md
@@ -17,6 +17,7 @@ Nussknacker versions
     * [#1335](https://github.com/TouK/nussknacker/pull/1335) Omitting zeros for sum aggregate to avoid unnecessary buckets 
     * [#1336](https://github.com/TouK/nussknacker/pull/1336) Aggregation metrics
 * [#1321](https://github.com/TouK/nussknacker/pull/1321) Exception handler accessible via custom node context, avro record encoding errors reported by exception handler
+* [#1343](https://github.com/TouK/nussknacker/pull/1359) Improvements in variable output validation
 
 0.3.0 
 ------------------------